+++ src/components/month_picker/month_picker.js
@@ -15,6 +15,7 @@ export default function MonthPicker({
   handleMonthChange,
   handleFocusedDate,
   rangeHover,
+  highlightToday,
 }) {
   const {
       date,
@@ -142,7 +143,10 @@ export default function MonthPicker({
       names.push("rmdp-disabled");
 
     if (names.includes("rmdp-disabled") && onlyShowInRangeDates) return;
-    if (isSameDate(today, dateObject, true)) names.push("rmdp-today");
+
+    if (isSameDate(today, dateObject, true) && highlightToday) {
+      names.push("rmdp-today");
+    }
 
     if (!onlyMonthPicker) {
       if (date.monthIndex === monthIndex) names.push("rmdp-selected");